Split transcriber implementation tests and TranscribeCommand tests to their own modules. Added mock implementation for audio conversion so that ffmpeg is no longer needed for the tests
